Moved CPU code out of core module
