+++ src/core/mod.rs
@@ -1,7 +1,4 @@
-pub mod errors;
 pub mod byte;
 pub mod double;
 pub mod memory;
-pub mod cpu;
 pub mod consts;
-pub mod instructions;
