+++ src/cpu/errors.rs
