+++ src/cpu/mod.rs
@@ -1,0 +1,3 @@
+pub mod errors;
+pub mod instructions;
+pub mod cpu;
